refactor(state): :recycle: state machine include refactoring
State machine functions should only be exposed by a single "state.h"
header file instead of using "simple.h".
