+++ lib/state/simple.c
@@ -1,6 +1,6 @@
 /**
- * @file simple_state.c
- * @brief Simple 9-state flight state machine implementation.
+ * @file simple.c
+ * @brief Simple 9-state flight state machine backend.
  *
  * Implements the flight state sequence:
  * IDLE -> ARMED -> BOOST -> BURNOUT -> APOGEE -> MAIN -> REDUNDANT -> LANDED / ERROR
@@ -8,53 +8,30 @@
  * State transitions are driven by sensor thresholds and timers.
  * Optionally integrates with the Kalman filter input filtering.
  *
+ * Only the implementation-specific pieces live here: the threshold set,
+ * the per-state transition logic and its timers.  The common lifecycle,
+ * update entry point and error handling are provided by the state core
+ * (state.c); this backend drives it through the helpers in
+ * state_internal.h.
+ *
  * Copyright (c) 2025-2026, Auxspace e.V.
  *
  * SPDX-License-Identifier: Apache-2.0
  */
 
 #include <math.h>
+#include <string.h>
 #include <zephyr/kernel.h>
 #include <zephyr/logging/log.h>
-#include <zephyr/spinlock.h>
 
 #include <aurora/lib/state/state.h>
+#include "state_internal.h"
 
 #ifndef M_PI
 #define M_PI ((double)3.1415926535)
 #endif
 
-/**
- * @brief Elevation of the configured up axis from horizontal (degrees).
- *
- * The orientation vector produced by imu_sensor_value_to_orientation()
- * already accounts for @c CONFIG_IMU_UP_AXIS_* by remapping the body
- * "up" axis to local Z, so pitch and yaw alone are sufficient to
- * recover the up-axis tilt:
- *   gz/|g| = cos(pitch) * cos(yaw)
- *
- * @return Elevation in degrees, clamped to [-90, 90].  +90 = up axis
- *         points to the sky, 0 = horizontal, -90 = inverted.
- */
-static inline double sm_orientation_elevation_deg(const double orientation[3])
-{
-	const double deg2rad = M_PI / 180.0;
-	double s = cos(orientation[1] * deg2rad) * cos(orientation[0] * deg2rad);
-	if (s > 1.0)  s = 1.0;
-	if (s < -1.0) s = -1.0;
-	return asin(s) * (180.0 / M_PI);
-}
-
-#if defined(CONFIG_AURORA_STATE_MACHINE_AUDIT)
-#include <aurora/lib/state/audit.h>
-#endif /* CONFIG_AURORA_STATE_MACHINE_AUDIT */
-
-#if defined(CONFIG_FILTER)
-#include <aurora/lib/filter.h>
-static struct filter filter;
-#endif /* CONFIG_FILTER */
-
-LOG_MODULE_REGISTER(simple_state, CONFIG_STATE_MACHINE_LOG_LEVEL);
+LOG_MODULE_DECLARE(state_machine, CONFIG_STATE_MACHINE_LOG_LEVEL);
 
 /*-----------------------------------------------------------
  * Prototypes
@@ -62,26 +39,6 @@ LOG_MODULE_REGISTER(simple_state, CONFIG_STATE_MACHINE_LOG_LEVEL);
 
 /** @brief Initialize all internal Zephyr timers used by the state machine. */
 static void init_timers(void);
-
-/** @brief Stop all running timers and clear the @ref running_timers flags. */
-static void stop_timers(void);
-
-/**
- * @brief Default error handler when no user callback is registered.
- *
- * @param reason Why the state machine entered SM_ERROR.
- * @param args Unused opaque argument (always NULL).
- * @return Always returns -EIO.
- */
-static int fallback_sm_error_handler(enum sm_error_reason reason, void *args);
-
-/**
- * @brief Transition to SM_ERROR and invoke the error callback.
- *
- * Acquires @ref err_lock, sets state to SM_ERROR, records @p reason and
- * calls the registered error callback (or the fallback) with it.
- */
-static void sm_do_error_handling(enum sm_error_reason reason);
 
 /**
  * @brief Check if acceleration and altitude meet ARMED->BOOST thresholds.
@@ -92,30 +49,10 @@ static void sm_do_error_handling(enum sm_error_reason reason);
  */
 static inline bool arm_to_boost_conditions_met(const struct sm_inputs *in);
 
-/**
- * @brief Core state machine update logic.
- *
- * Evaluates sensor inputs against thresholds and timers to determine
- * state transitions.  Called by sm_update() with additional bookkeeping.
- *
- * @param in                Pointer to the current sensor input values.
- * @param previous_altitude Altitude from the previous update cycle (m).
- */
-static inline void _sm_update(const struct sm_inputs *in,
-			      double previous_altitude);
-
 /*-----------------------------------------------------------
  * Internal State
  *----------------------------------------------------------*/
-static enum sm_state current_state = SM_IDLE; /**< Active flight state. */
 static struct sm_thresholds th; /**< Loaded threshold configuration. */
-static struct sm_inputs last_inputs; /**< Last inputs evaluated by _sm_update(). */
-
-static struct k_spinlock err_lock; /**< Spinlock protecting error callback invocation. */
-static struct sm_error_handling_args err_hdl = {
-	.cb = &fallback_sm_error_handler,
-	.args = NULL,
-};
 
 static struct k_timer dt_ab; /**< Duration timer for ARMED->BOOST assertion. */
 static struct k_timer dt_l;  /**< Duration timer for landing velocity assertion. */
@@ -145,16 +82,26 @@ static int running_timers[NUM_TIMERS] = {0}; /**< Per-timer running flag. */
  */
 #define TIMER_EXPIRED(tmr) (k_timer_status_get(tmr) > 0)
 
-#if defined(CONFIG_AURORA_STATE_MACHINE_AUDIT)
-#define SM_TRANSITION(new_state) do {				\
-	sm_audit_transition(current_state, (new_state));	\
-	current_state = (new_state);				\
-} while (0)
-#define SM_EVENT(msg) sm_audit_event(current_state, (msg))
-#else
-#define SM_TRANSITION(new_state) do { current_state = (new_state); } while (0)
-#define SM_EVENT(msg)
-#endif /* CONFIG_AURORA_STATE_MACHINE_AUDIT */
+/**
+ * @brief Elevation of the configured up axis from horizontal (degrees).
+ *
+ * The orientation vector produced by imu_sensor_value_to_orientation()
+ * already accounts for @c CONFIG_IMU_UP_AXIS_* by remapping the body
+ * "up" axis to local Z, so pitch and yaw alone are sufficient to
+ * recover the up-axis tilt:
+ *   gz/|g| = cos(pitch) * cos(yaw)
+ *
+ * @return Elevation in degrees, clamped to [-90, 90].  +90 = up axis
+ *         points to the sky, 0 = horizontal, -90 = inverted.
+ */
+static inline double sm_orientation_elevation_deg(const double orientation[3])
+{
+	const double deg2rad = M_PI / 180.0;
+	double s = cos(orientation[1] * deg2rad) * cos(orientation[0] * deg2rad);
+	if (s > 1.0)  s = 1.0;
+	if (s < -1.0) s = -1.0;
+	return asin(s) * (180.0 / M_PI);
+}
 
 static void init_timers(void)
 {
@@ -168,7 +115,8 @@ static void init_timers(void)
 	memset(running_timers, 0, sizeof(running_timers));
 }
 
-static void stop_timers(void)
+/* sm_backend_stop_timers – see state_internal.h */
+void sm_backend_stop_timers(void)
 {
 	k_timer_stop(&dt_ab);
 	k_timer_stop(&dt_l);
@@ -180,113 +128,46 @@ static void stop_timers(void)
 	memset(running_timers, 0, sizeof(running_timers));
 }
 
-static int fallback_sm_error_handler(enum sm_error_reason reason, void *args)
-{
-	ARG_UNUSED(args);
-	LOG_ERR("State Machine encountered an unrecoverable error (%s)",
-		sm_error_reason_str(reason));
-
-	return -EIO;
-}
-
-/* Why the machine is (or last was) in SM_ERROR.  Recorded by
- * sm_do_error_handling() so re-invocations from the SM_ERROR state keep
- * reporting the original cause to the callback.
- */
-static enum sm_error_reason err_reason = SM_ERR_UNKNOWN;
-
-static void sm_do_error_handling(enum sm_error_reason reason)
-{
-	int ret;
-	k_spinlock_key_t key = k_spin_lock(&err_lock);
-
-	if (current_state != SM_ERROR) {
-		err_reason = reason;
-		SM_TRANSITION(SM_ERROR);
-	}
-
-	if (err_hdl.cb == NULL) {
-		LOG_ERR("No fallback handler defined for state machine errors!");
-		goto out;
-	}
-
-	ret = err_hdl.cb(err_reason, err_hdl.args);
-	if (ret == 0) {
-		SM_EVENT("error mitigated, returning to IDLE");
-		stop_timers();
-		err_reason = SM_ERR_UNKNOWN;
-		SM_TRANSITION(SM_IDLE);
-	} else {
-		LOG_ERR("State Machine error handler failed with code %d", ret);
-	}
-
-out:
-	k_spin_unlock(&err_lock, key);
-}
-
-/*-----------------------------------------------------------
- * Initialization / Deinitialization
- *----------------------------------------------------------*/
-
-/* sm_init – see state.h */
-void sm_init(const struct sm_thresholds *cfg,
-			 struct sm_error_handling_args *sm_err_hdl)
-{
-#if defined(CONFIG_FILTER)
-	int ret = filter_init(&filter);
-	if (ret) {
-		LOG_ERR("Could not initialize filter (%d).", ret);
-		return;
-	}
-#endif /* CONFIG_FILTER */
-
-	th = *(struct sm_thresholds *)cfg;
-	init_timers();
-	current_state = SM_IDLE;
-	err_reason = SM_ERR_UNKNOWN;
-	SM_EVENT("state machine initialized");
-
-	if (sm_err_hdl != NULL) {
-		err_hdl.cb = sm_err_hdl->cb;
-		err_hdl.args = sm_err_hdl->args;
-	}
-}
-
-/* sm_deinit – see state.h */
-void sm_deinit(void)
-{
-	memset(&th, 0, sizeof(th));
-	memset(&last_inputs, 0, sizeof(last_inputs));
-	stop_timers();
-	SM_EVENT("state machine reset");
-	current_state = SM_IDLE;
-	err_reason = SM_ERR_UNKNOWN;
-}
-
-/*-----------------------------------------------------------
- * Helper for ARM -> BOOST conditions
- *----------------------------------------------------------*/
 static inline bool arm_to_boost_conditions_met(const struct sm_inputs *in)
 {
 	return (in->acceleration >= th.T_AB &&
-			in->altitude >= th.T_H);
+		in->altitude >= th.T_H);
 }
 
 /*-----------------------------------------------------------
- * State Machine Update
+ * Initialization / Deinitialization (see state_internal.h)
  *----------------------------------------------------------*/
-static inline void _sm_update(const struct sm_inputs *in,
-			      double previous_altitude)
+
+/* sm_backend_init – see state_internal.h */
+void sm_backend_init(const struct sm_thresholds *cfg)
+{
+	th = *cfg;
+	init_timers();
+}
+
+/* sm_backend_deinit – see state_internal.h */
+void sm_backend_deinit(void)
+{
+	memset(&th, 0, sizeof(th));
+	sm_backend_stop_timers();
+}
+
+/*-----------------------------------------------------------
+ * State Machine Update (see state_internal.h)
+ *----------------------------------------------------------*/
+
+/* sm_backend_step – see state_internal.h */
+void sm_backend_step(const struct sm_inputs *in, double previous_altitude)
 {
 	static int n_oi;
 
 	/* go to IDLE if disarmed */
-	if (!in->armed && current_state != SM_IDLE) {
-		stop_timers();
-		SM_TRANSITION(SM_IDLE);
+	if (!in->armed && sm_get_state() != SM_IDLE) {
+		sm_backend_stop_timers();
+		sm_transition(SM_IDLE);
 	}
 
-	switch (current_state)
+	switch (sm_get_state())
 	{
 
 	/*-----------------------------------------------------------
@@ -296,7 +177,7 @@ static inline void _sm_update(const struct sm_inputs *in,
 		n_oi = 0;
 		if (in->armed && sm_orientation_elevation_deg(in->orientation) >= th.T_OA) {
 			if (in->log_ready) {
-				SM_TRANSITION(SM_ARMED);
+				sm_transition(SM_ARMED);
 			} else {
 				/* Arm conditions are met but the flight log is
 				 * offline.  Refuse to arm so the vehicle never
@@ -305,7 +186,7 @@ static inline void _sm_update(const struct sm_inputs *in,
 				 * operator gets an unmissable field indication
 				 * (LED/buzzer via the app error callback).
 				 */
-				SM_EVENT("arm refused: flight log offline");
+				sm_event("arm refused: flight log offline");
 				sm_do_error_handling(SM_ERR_LOG_OFFLINE);
 			}
 		}
@@ -326,7 +207,7 @@ static inline void _sm_update(const struct sm_inputs *in,
 		if (!in->log_ready) {
 			running_timers[TIMER_DT_AB] = 0;
 			k_timer_stop(&dt_ab);
-			SM_EVENT("arm aborted: flight log offline");
+			sm_event("arm aborted: flight log offline");
 			sm_do_error_handling(SM_ERR_LOG_OFFLINE);
 			break;
 		}
@@ -338,8 +219,8 @@ static inline void _sm_update(const struct sm_inputs *in,
 			/* Go back to IDLE if orientation is bad */
 			running_timers[TIMER_DT_AB] = 0;
 			k_timer_stop(&dt_ab);
-			SM_EVENT("orientation below threshold");
-			SM_TRANSITION(SM_IDLE);
+			sm_event("orientation below threshold");
+			sm_transition(SM_IDLE);
 			break;
 		}
 		n_oi = 0;
@@ -357,9 +238,9 @@ static inline void _sm_update(const struct sm_inputs *in,
 			/* At this point, conditions are met. Is the timer done as well? */
 			if (TIMER_EXPIRED(&dt_ab)) {
 				n_oi = 0;
-				SM_EVENT("orientation, altitude and timing threshold reached");
+				sm_event("orientation, altitude and timing threshold reached");
 				/* Congrats! BOOST detected! */
-				SM_TRANSITION(SM_BOOST);
+				sm_transition(SM_BOOST);
 				k_timer_stop(&dt_ab);
 				running_timers[TIMER_DT_AB] = 0;
 			}
@@ -371,7 +252,7 @@ static inline void _sm_update(const struct sm_inputs *in,
 		if (!arm_to_boost_conditions_met(in))
 			break;
 
-		SM_EVENT("orientation and altitude threshold reached");
+		sm_event("orientation and altitude threshold reached");
 		/* Conditions are met, so start the timer. */
 		k_timer_start(&dt_ab, K_MSEC(th.DT_AB), K_NO_WAIT);
 		running_timers[TIMER_DT_AB] = 1;
@@ -382,7 +263,7 @@ static inline void _sm_update(const struct sm_inputs *in,
 	*----------------------------------------------------------*/
 	case SM_BOOST:
 		if (in->acceleration < th.T_BB) {
-			SM_TRANSITION(SM_BURNOUT);
+			sm_transition(SM_BURNOUT);
 		}
 		break;
 
@@ -391,14 +272,14 @@ static inline void _sm_update(const struct sm_inputs *in,
 	*----------------------------------------------------------*/
 	case SM_BURNOUT:
 #if defined(CONFIG_FILTER)
-		if (filter_detect_apogee(&filter) == 1) {
+		if (sm_filter_detect_apogee() == 1) {
 			k_timer_start(&to_a, K_MSEC(th.TO_A), K_NO_WAIT);
-			SM_TRANSITION(SM_APOGEE);
+			sm_transition(SM_APOGEE);
 		}
 #else
 		if (in->velocity <= 0.0 && in->altitude < previous_altitude) {
 			k_timer_start(&to_a, K_MSEC(th.TO_A), K_NO_WAIT);
-			SM_TRANSITION(SM_APOGEE);
+			sm_transition(SM_APOGEE);
 		}
 #endif /* CONFIG_FILTER */
 		break;
@@ -410,11 +291,11 @@ static inline void _sm_update(const struct sm_inputs *in,
 		if (in->altitude < th.T_M) {
 			k_timer_stop(&to_a);
 			k_timer_start(&to_m, K_MSEC(th.TO_M), K_NO_WAIT);
-			SM_TRANSITION(SM_MAIN);
+			sm_transition(SM_MAIN);
 		} else if (TIMER_EXPIRED(&to_a)) {
 			/* Timeout expired, abort to ERROR */
 			k_timer_stop(&to_a);
-			SM_EVENT("apogee timeout expired");
+			sm_event("apogee timeout expired");
 			sm_do_error_handling(SM_ERR_APOGEE_TIMEOUT);
 		}
 		break;
@@ -426,7 +307,7 @@ static inline void _sm_update(const struct sm_inputs *in,
 		if (TIMER_EXPIRED(&to_m)) {
 			k_timer_stop(&to_m);
 			k_timer_start(&to_r, K_MSEC(th.TO_R), K_NO_WAIT);
-			SM_TRANSITION(SM_REDUNDANT);
+			sm_transition(SM_REDUNDANT);
 		}
 		break;
 
@@ -446,7 +327,7 @@ static inline void _sm_update(const struct sm_inputs *in,
 			/* At this point, conditions are met. Is the timer done as well? */
 			if (TIMER_EXPIRED(&dt_l)) {
 				/* Congrats! LANDING detected! */
-				SM_TRANSITION(SM_LANDED);
+				sm_transition(SM_LANDED);
 				k_timer_stop(&dt_l);
 				running_timers[TIMER_DT_L] = 0;
 				break;
@@ -470,7 +351,7 @@ _check_timeout:
 			k_timer_stop(&dt_l);
 			running_timers[TIMER_DT_L] = 0;
 			k_timer_stop(&to_r);
-			SM_EVENT("redundant timeout expired");
+			sm_event("redundant timeout expired");
 			sm_do_error_handling(SM_ERR_REDUNDANT_TIMEOUT);
 		}
 		break;
@@ -479,7 +360,7 @@ _check_timeout:
 	*----------------------------------------------------------*/
 	case SM_ERROR:
 		/* Try to fix the error; re-report the original cause. */
-		sm_do_error_handling(err_reason);
+		sm_error_retry();
 		break;
 
 	/*-----------------------------------------------------------
@@ -491,77 +372,14 @@ _check_timeout:
 	}
 }
 
-/* sm_update – see state.h */
-void sm_update(const struct sm_inputs *inputs)
-{
-	static double previous_altitude = 0.0;
-
-#if defined(CONFIG_FILTER)
-	static uint64_t last_time_ns = 0;
-	struct sm_inputs filtered_inputs;
-
-	uint64_t current_time_ns = k_ticks_to_ns_floor64(k_uptime_ticks());
-
-	if (last_time_ns != 0) {
-		filter_predict(&filter, current_time_ns - last_time_ns,
-			       inputs->accel_vert);
-		filter_update(&filter, inputs->altitude);
-	}
-	last_time_ns = current_time_ns;
-
-	filtered_inputs = *inputs;
-	filtered_inputs.altitude = filter.state[0];
-	filtered_inputs.velocity = filter.state[1];
-
-	_sm_update(&filtered_inputs, previous_altitude);
-	previous_altitude = filtered_inputs.altitude;
-	last_inputs = filtered_inputs;
-#else
-	_sm_update(inputs, previous_altitude);
-	previous_altitude = inputs->altitude;
-	last_inputs = *inputs;
-#endif /* CONFIG_FILTER */
-}
-
-/* sm_update_force – see state.h */
-void sm_update_force(enum sm_state transition_to)
-{
-	SM_TRANSITION(transition_to);
-}
-
-
 /*-----------------------------------------------------------
- * Getter
+ * Getters
  *----------------------------------------------------------*/
-
-/* sm_get_state – see state.h */
-enum sm_state sm_get_state(void)
-{
-	return current_state;
-}
 
 /* sm_get_type - see state.h */
 enum sm_type sm_get_type(void)
 {
 	return SM_TYPE_SIMPLE;
-}
-
-/* sm_get_inputs – see state.h */
-void sm_get_inputs(struct sm_inputs *out)
-{
-	*out = last_inputs;
-}
-
-/* sm_error_reason_str – see state.h */
-const char *sm_error_reason_str(enum sm_error_reason reason)
-{
-	switch (reason) {
-	case SM_ERR_UNKNOWN:		return "UNKNOWN";
-	case SM_ERR_LOG_OFFLINE:	return "LOG_OFFLINE";
-	case SM_ERR_APOGEE_TIMEOUT:	return "APOGEE_TIMEOUT";
-	case SM_ERR_REDUNDANT_TIMEOUT:	return "REDUNDANT_TIMEOUT";
-	default:			return "UNKNOWN";
-	}
 }
 
 /* sm_state_str – see simple.h */
